drop support for DBAL 2
