+++ tests/DAMA/DoctrineTestBundle/Doctrine/DBAL/MockDriver.php
@@ -2,14 +2,68 @@
 
 namespace Tests\DAMA\DoctrineTestBundle\Doctrine\DBAL;
 
-if (interface_exists(\Doctrine\DBAL\Driver\ExceptionConverterDriver::class)) {
-    // dbal v2
-    class MockDriver extends MockDriverV2
+use Doctrine\DBAL\Connection;
+use Doctrine\DBAL\Driver;
+use Doctrine\DBAL\Driver\API\ExceptionConverter;
+use Doctrine\DBAL\Platforms\AbstractPlatform;
+use Doctrine\DBAL\Schema\AbstractSchemaManager;
+use PHPUnit\Framework\MockObject\Generator;
+
+class MockDriver implements Driver
+{
+    private function getMock(string $class)
     {
+        return (new Generator())->getMock(
+            $class,
+            [],
+            [],
+            '',
+            false
+        );
     }
-} else {
-    // dbal v3
-    class MockDriver extends MockDriverV3
+
+    /**
+     * {@inheritdoc}
+     */
+    public function connect(array $params): \Doctrine\DBAL\Driver\Connection
     {
+        return $this->getMock(Driver\Connection::class);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getDatabasePlatform(): AbstractPlatform
+    {
+        return $this->getMock(AbstractPlatform::class);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getSchemaManager(Connection $conn, AbstractPlatform $platform): AbstractSchemaManager
+    {
+        return $this->getMock(AbstractSchemaManager::class);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getName(): string
+    {
+        return 'mock';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getDatabase(Connection $conn): string
+    {
+        return 'mock';
+    }
+
+    public function getExceptionConverter(): ExceptionConverter
+    {
+        return $this->getMock(ExceptionConverter::class);
     }
 }
